[fix] one report loads now
