+++ reports/local-income-expense.js
@@ -3,54 +3,66 @@
  *
  *
  */
-
+const fs = require("fs");
 const path = require("path");
-
-const ABApplication = require(path.join(
-   "..",
-   "classes",
-   "platform",
-   "ABApplication"
-));
 
 module.exports = {
    // GET: /template/localIncomeExpense
    // get the local and expense income and calculate the sums
-   getData: function(req, res) {
+   // ! getData: function(req, res) {
+   prepareData: async (AB, { rc, fiscalPeriod }) => {
       // get our passed params
       //console.log("params -------------->", req);
-      let rc = req.query.rc ? req.query.rc : undefined;
-      let fyper = req.query.fyper ? req.query.fyper : undefined;
-      // get the users preferred language
-      let languageCode = req.user.data.languageCode
-         ? req.user.data.languageCode
-         : "en";
+      rc = rc ? rc : undefined;
+      fiscalPeriod = fiscalPeriod ? fiscalPeriod : undefined;
 
-      if (req.query.languageCode) {
-         languageCode = req.query.languageCode;
+      const ids = {
+         myRCsQueryId: "241a977c-7748-420d-9dcb-eff53e66a43f",
+         balanceObjId: "bb9aaf02-3265-4b8c-9d9a-c0b447c2d804",
+         fiscalMonthObjId: "1d63c6ac-011a-4ffd-ae15-97e5e43f2b3f",
+      };
+
+      // TODO make sure there's a language code
+      // // get the users preferred language
+      // let languageCode = req.user.data.languageCode
+      //    ? req.user.data.languageCode
+      //    : "en";
+
+      if (AB.query && AB.query.languageCode) {
+         console.log(
+            "languageCode --------------~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~>",
+            AB.query.languageCode
+         );
+         languageCode = AB.query.languageCode;
+      }
+      if (AB.user && AB.user.data) {
+         console.log(
+            "User --------------~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~>",
+            AB.user.data
+         );
       }
 
-      if (languageCode == "zh-hans") {
-         languageCode = "zh";
-      }
+      // if (languageCode == "zh-hans") {
+      var languageCode = "zh";
+      // }
 
       //console.log("language ------->", languageCode);
 
       // Our data object
-      let data = {
+      var data = {
          title: {
             en: "Local Income vs Expense",
-            zh: "本地收入VS 支出"
+            zh: "本地收入VS 支出",
          },
          rc: rc,
          languageCode: languageCode,
          total: {
             en: "Total",
-            zh: "总额"
+            zh: "总额",
          },
          category: {
             en: "Category",
-            zh: "类别"
+            zh: "类别",
          },
          categories: [
             {
@@ -58,99 +70,99 @@ module.exports = {
                type: "Local Income",
                translation: {
                   en: "Local Income ",
-                  zh: "本地收入"
+                  zh: "本地收入",
                },
                sub: [
                   {
                      id: 41113,
                      translation: {
                         en: "General Contribution Local From Ch",
-                        zh: "本地收到给事工的捐款-从国内收到"
-                     }
+                        zh: "本地收到给事工的捐款-从国内收到",
+                     },
                   },
                   {
                      id: 41114,
                      translation: {
                         en: "General Contribution Local From Oversea",
-                        zh: "本地收到给事工的捐款-收到海外的汇款"
-                     }
-                  }
-               ]
+                        zh: "本地收到给事工的捐款-收到海外的汇款",
+                     },
+                  },
+               ],
             },
             {
                parent: 7,
                type: "Expenses",
                translation: {
                   en: "Expenses ",
-                  zh: "支出"
+                  zh: "支出",
                },
                sub: [
                   {
                      id: 71,
                      translation: {
                         en: "Personnel expenses",
-                        zh: "工资/福利"
-                     }
+                        zh: "工资/福利",
+                     },
                   },
                   {
                      id: 72,
                      translation: {
                         en: "Conferences and meetings",
-                        zh: "大会和会议费用"
-                     }
+                        zh: "大会和会议费用",
+                     },
                   },
                   {
                      id: 75,
                      translation: {
                         en: "Travel and transportation",
-                        zh: "差旅费"
-                     }
+                        zh: "差旅费",
+                     },
                   },
                   {
                      id: 81,
                      translation: {
                         en: "Supplies and non-capitalized equipment",
-                        zh: "设备及维修保养"
-                     }
+                        zh: "设备及维修保养",
+                     },
                   },
                   {
                      id: 82,
                      translation: {
                         en: "Communications",
-                        zh: "电话和通信"
-                     }
+                        zh: "电话和通信",
+                     },
                   },
                   {
                      id: 84,
                      translation: {
                         en: "Professional services",
-                        zh: "专业费用"
-                     }
+                        zh: "专业费用",
+                     },
                   },
                   {
                      id: 86,
                      translation: {
                         en: "Capital expenses",
-                        zh: "固定资产支出"
-                     }
+                        zh: "固定资产支出",
+                     },
                   },
                   {
                      id: 87,
                      translation: {
                         en: "Facilities",
-                        zh: "设施费用"
-                     }
+                        zh: "设施费用",
+                     },
                   },
                   {
                      id: 89,
                      translation: {
                         en: "Other expenses",
-                        zh: "其他费用"
-                     }
-                  }
-               ]
-            }
-         ]
+                        zh: "其他费用",
+                     },
+                  },
+               ],
+            },
+         ],
       };
 
       function accountInCategory(account, category) {
@@ -178,31 +190,32 @@ module.exports = {
          }
       }
 
-      let myRCs = ABSystemObject.getApplication().queries(
-         (o) => o.id == "241a977c-7748-420d-9dcb-eff53e66a43f"
-      )[0];
+      // let myRCs = ABSystemObject.getApplication().queries(
+      //    (o) => o.id == "241a977c-7748-420d-9dcb-eff53e66a43f"
+      // )[0];
+      const myRCs = AB.queryByID(ids.myRCsQueryId)?.model();
 
-      //console.log("myRCs ----------------->", myRCs);
+      console.log("myRCs ----------------->", myRCs);
 
-      myRCs
+      return myRCs
          .queryFind(
             {
                where: {
                   glue: "and",
-                  rules: []
-               }
+                  rules: [],
+               },
             },
-            req.user.data
+            AB.user.data
          )
          .then((rcs) => {
-            //console.log("My Team RCs ---------------->", rcs);
+            console.log("My Team RCs ---------------->", rcs);
 
             let rcOptions = [];
             rcs.forEach((rc) => {
                rcOptions.push(rc["BASE_OBJECT.RC Name"]);
             });
 
-            data.rcOptions = rcOptions.sort(function(a, b) {
+            data.rcOptions = rcOptions.sort(function (a, b) {
                return a.toLowerCase().localeCompare(b.toLowerCase());
             });
 
@@ -211,11 +224,12 @@ module.exports = {
                data.rc = rc;
             }
 
-            let fiscalMonthObj = ABSystemObject.getApplication().objects(
-               (o) => o.id == "1d63c6ac-011a-4ffd-ae15-97e5e43f2b3f"
-            )[0];
+            const fiscalMonthObj = AB.objectByID(ids.fiscalMonthObjId).model();
+            // let fiscalMonthObj = ABSystemObject.getApplication().objects(
+            //    (o) => o.id == "1d63c6ac-011a-4ffd-ae15-97e5e43f2b3f"
+            // )[0];
 
-            fiscalMonthObj
+            return fiscalMonthObj
                .modelAPI()
                .findAll({
                   where: {
@@ -224,24 +238,28 @@ module.exports = {
                         {
                            key: "Status",
                            rule: "equals",
-                           value: "1592549786113"
-                        }
-                     ]
+                           value: "1592549786113",
+                        },
+                     ],
                   },
                   populate: false,
                   sort: [
                      {
                         key: "49d6fabe-46b1-4306-be61-1b27764c3b1a",
-                        dir: "DESC"
-                     }
+                        dir: "DESC",
+                     },
                   ],
-                  limit: 12
+                  limit: 12,
                })
                .then((records) => {
-                  //console.log("Fiscal Month Records ------------------>", records);
+                  console.log(
+                     "Fiscal Month Records ------------------>",
+                     records
+                  );
                   let fiscalMonthsArray = records;
-                  data.fyper = fyper || fiscalMonthsArray[0]["FY Per"];
-                  let fyperOptions = [];
+                  data.fiscalPeriod =
+                     fiscalPeriod || fiscalMonthsArray[0]["FY Per"];
+                  let fiscalPeriodOptions = [];
                   let i = 0;
                   let currIndex = 0;
                   fiscalMonthsArray.forEach((fp) => {
@@ -251,29 +269,33 @@ module.exports = {
                      var prettyDate =
                         year + "/" + (month > 9 ? month : "0" + month);
                      var option = { id: fp["FY Per"], label: prettyDate };
-                     if (fyper == fp["FY Per"]) {
+                     if (fiscalPeriod == fp["FY Per"]) {
                         option.selected = true;
                         currIndex = i;
                      }
-                     fyperOptions.push(option);
+                     fiscalPeriodOptions.push(option);
                      i++;
                   });
-                  data.fyperOptions = fyperOptions;
+                  data.fiscalPeriodOptions = fiscalPeriodOptions;
                   var dateObj = new Date(fiscalMonthsArray[currIndex]["End"]);
                   var month = dateObj.getUTCMonth() + 1; //months from 1-12
                   var year = dateObj.getUTCFullYear();
-                  data.fyperend =
+                  data.fiscalPeriodend =
                      year + "/" + (month > 9 ? month : "0" + month);
                   let startYear = year;
                   if (month < 7) {
                      startYear = year - 1;
                   }
-                  data.fyperstart = startYear + "/07";
+                  data.fiscalPeriodstart = startYear + "/07";
 
-                  //console.log("Fiscal Month picked from query param -->", data.fyper);
-                  let balanceObj = ABSystemObject.getApplication().objects(
-                     (o) => o.id == "bb9aaf02-3265-4b8c-9d9a-c0b447c2d804"
-                  )[0];
+                  console.log(
+                     "Fiscal Month picked from query param -->",
+                     data.fiscalPeriod
+                  );
+                  const balanceObj = AB.objectByID(ids.balanceObjId).model();
+                  // let balanceObj = ABSystemObject.getApplication().objects(
+                  //    (o) => o.id == "bb9aaf02-3265-4b8c-9d9a-c0b447c2d804"
+                  // )[0];
 
                   balanceObj
                      .modelAPI()
@@ -284,19 +306,19 @@ module.exports = {
                               {
                                  key: "RC Code",
                                  rule: "equals",
-                                 value: rc
+                                 value: rc,
                               },
                               {
                                  key: "FY Period",
                                  rule: "equals",
-                                 value: data.fyper
-                              }
-                           ]
+                                 value: data.fiscalPeriod,
+                              },
+                           ],
                         },
-                        populate: false
+                        populate: false,
                      })
                      .then((records) => {
-                        //console.log(records);
+                        console.log(records);
 
                         data.categories.forEach((cat) => {
                            let catSum = 0;
@@ -312,12 +334,22 @@ module.exports = {
                               100
                         );
 
-                        res.view(
-                           "app_builder/template/localIncomeExpense", // .ejs
-                           data
+                        console.log(
+                           "data is outputting: -> -> -> -> -> -> -> -> ->"
                         );
+                        return data;
+                        // res.view(
+                        //    "app_builder/template/localIncomeExpense", // .ejs
+                        //    data
+                        // );
                      });
                });
          });
-   }
+   },
+   template: () => {
+      return fs.readFileSync(
+         path.join(__dirname, "templates", "local-income-expense.ejs"),
+         "utf8"
+      );
+   },
 };
